Cast argument to double in Sonar rule.

+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/EbicScore.java
@@ -130,7 +130,7 @@ public class EbicScore implements Score {
     }
 
     public static double getP(int pn, int m0, double lambda) {
-        return 2 - pow((1 + (exp(-(lambda - 1) / 2.)) * sqrt(lambda)), pn - m0);
+        return 2.0 - pow((1 + (exp(-(lambda - 1) / 2.)) * sqrt(lambda)), (double) pn - m0);
     }
 
     /**
